demodadsb/adsbdemod: Prevent integer overflow in ADS-B sample rate calculations
Coverity reported integer overflow before widening when calculating the
sample rate in both ADSBDemodSink and ADSBDemodSinkWorker.

Promote ADS_B_BITS_PER_SECOND to qint64 before multiplication so the
calculations are performed using 64-bit arithmetic.

+++ plugins/channelrx/demodadsb/adsbdemodsinkworker.cpp
@@ -693,7 +693,7 @@ void ADSBDemodSinkWorker::handleInputMessages()
 
 QDateTime ADSBDemodSinkWorker::rxDateTime(int firstIdx, int readBuffer) const
 {
-    const qint64 samplesPerSecondMSec = ADS_B_BITS_PER_SECOND * m_settings.m_samplesPerBit / 1000;
+    const qint64 samplesPerSecondMSec = qint64(ADS_B_BITS_PER_SECOND) * m_settings.m_samplesPerBit / 1000;
     const qint64 offsetMSec = (firstIdx - m_sink->m_samplesPerFrame - 1) / samplesPerSecondMSec;
     return m_sink->m_bufferFirstSampleDateTime[readBuffer].addMSecs(offsetMSec);
 }
